Move net segment splitter classes to core library
